pyproject.toml
[tool.poetry]
name="analyticalplatform"
version="0.1.0"
description="Analytical Platform"
readme="README.md"
authors=[]

[tool.poetry.dependencies]
python="^3.12"
django = "4.2.10"
djangorestframework = "3.14.0"
requests = "2.31.0"
django-cors-headers = "4.2.0"
django-filter = "^24.2"
celery = "5.3.6"
gunicorn = "21.2.0"
psycopg2-binary = "^2.9.9"
django-celery-beat = "2.5.0"
redis = "^5.0.3"
openpyxl = "^3.1.2"
tls-client = "^1.0.1"
typing-extensions = "^4.11.0"
django-silk = "^5.1.0"
pytest = "^8.1.1"
pytest-django = "^4.8.0"

[build-system]
requires=["poetry-core"]
build-backend="poetry.core.masonry.api"

[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "analyticalplatform.settings"
python_files = ["test_*.py", "*_test.py", "testing/python/*.py"]
